Revert "feat: add i18next plugin for gallery (WIP)"
This reverts commit 8b42539e6bd6a35e6fbb4d0b8ec50169dfb7a562.
