+++ src/plugins/i18next/lg-i18next.ts
@@ -1,8 +1,5 @@
 import { LightGallery } from '../../lightgallery';
-import i18next from 'i18next';
-import languageDetector from 'i18next-browser-languagedetector';
-import french from './fr';
-import english from './en';
+import TranslateService from './service';
 
 /**
  * Creates the i18next plugin.
@@ -14,36 +11,9 @@ export default class I18Next {
     constructor(instance: LightGallery) {
         this.core = instance;
 
-        this.init();
+        // Translation i18next initialization
+        TranslateService.init();
 
         return this;
-    }
-
-    private async init() {
-        const options = {
-            order: ['navigator'],
-        };
-
-        await i18next.use(languageDetector).init({
-            detection: options,
-            resources: {
-                en: {
-                    translation: english,
-                },
-                fr: {
-                    translation: french,
-                },
-            },
-        });
-
-        this.core.settings.strings.closeGallery = this.translate(
-            'closeGallery',
-        );
-
-        // How to refresh lightGallery after all translations?
-    }
-
-    private translate(key: string): string {
-        return i18next.t(key);
     }
 }
