+++ src/plugins/i18next/model.ts
@@ -1,0 +1,3 @@
+export interface ITranslate {
+    [key: string]: string;
+}
